+++ bin/git_changelogger.dart
@@ -7,7 +7,7 @@
 
 import 'dart:io';
 
-import 'package:changelog_cli/src/command_runner.dart';
+import 'package:git_changelogger/src/command_runner.dart';
 
 Future<void> main(List<String> args) async {
   await _flushThenExit(await ChangelogCliCommandRunner().run(args));
@@ -20,6 +20,5 @@ Future<void> main(List<String> args) async {
 /// exited already. This is useful to prevent Future chains from proceeding
 /// after you've decided to exit.
 Future<void> _flushThenExit(int status) {
-  return Future.wait<void>([stdout.close(), stderr.close()])
-      .then<void>((_) => exit(status));
+  return Future.wait<void>([stdout.close(), stderr.close()]).then<void>((_) => exit(status));
 }