+++ lib/src/commands/sample_command.dart
@@ -9,7 +9,7 @@ import 'package:args/command_runner.dart';
 import 'package:mason_logger/mason_logger.dart';
 
 /// {@template sample_command}
-/// `changelog_cli sample`
+/// `git_changelogger sample`
 /// A [Command] to exemplify a sub command
 /// {@endtemplate}
 class SampleCommand extends Command<int> {