+++ lib/src/commands/update_command.dart
@@ -6,8 +6,8 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:args/command_runner.dart';
-import 'package:changelog_cli/src/command_runner.dart';
-import 'package:changelog_cli/src/version.dart';
+import 'package:git_changelogger/src/command_runner.dart';
+import 'package:git_changelogger/src/version.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:pub_updater/pub_updater.dart';